+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Introduced String.graphemeCount extension for counting perceived emojis
+
 ### Changed
 
 - Use password type keyboard to enter recovery key